Added HealthDisplaySystem and Heart component

+++ src/settings.rs
@@ -45,8 +45,8 @@ pub struct SettingsPlayer {
     pub quick_turnaround:           SettingsPlayerQuickTurnaround,
     pub air_quick_turnaround:       SettingsPlayerQuickTurnaround,
     pub decrease_x_velocity_in_air: bool,
-    pub health:                     f32,
-    pub damage:                     f32,
+    pub health:                     u32,
+    pub damage:                     u32,
 }
 
 #[derive(Clone)]
@@ -78,9 +78,9 @@ pub struct SettingsEnemies {
 
 #[derive(Clone, Deserialize)]
 pub struct SettingsEnemy {
-    pub health:        f32,
-    pub damage:        f32,
-    pub reward:        f32,
+    pub health:        u32,
+    pub damage:        u32,
+    pub reward:        u32,
     pub knockback:     Vector,
     pub max_velocity:  (Option<f32>, Option<f32>),
     pub decr_velocity: Vector,
